Revert "Flatten the scanner"
This reverts commit 896877bcee6c719a8c95c2145a17b7c25ac456e1.
